removed silly uses of 'long' (now int, or qbyte *, depending on the way it was used), and changed memheader sentinels to work on processors which can not handle unaligned memory access (Sparc for example)
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@1734 d7cf8633-e32d-0410-b094-e92efae38249

+++ gl_backend.c
@@ -861,7 +861,7 @@ void R_Mesh_AddTransparent(void)
 		if (center < 0.0f)
 			center = 0.0f;
 		center += 8388608.0f;
-		i = *((long *)&center) & 0x7FFFFF;
+		i = *((int *)&center) & 0x7FFFFF;
 		i = min(i, (TRANSDEPTHRES - 1));
 #endif
 		tri->next = buf_sorttranstri_list[i];
